Undef TYPEOF macro set by Emscripten's Freetype
Emscripten's Wasm port of Freetype uses an older upstream source
that sets a macro named `TYPEOF` in its headers. This conflicts
with R's C API `TYPEOF` macro.

Later versions of Freetype>=2.6.2 renamed this macro to `FT_TYPEOF`.

We don't need the Freetype macro, so if we're compiling under
Emscripten explicitly `undef` it so that it can then be used by R.